updated first page of create a new space info in a reducer
